refactor(mcp): split server registrations
